Conflicts:
	QredoSDK/QredoRendezvous.m

+++ QredoSDK/include/QredoTypes.h
@@ -7,7 +7,7 @@
 
 #import "QredoQUID.h"
 
-typedef NSUInteger QredoAccessLevel; // for now just an integer, but probably needs enum values
+typedef int32_t QredoAccessLevel; // for now just an integer, but probably needs enum values
 
 typedef NS_ENUM(NSUInteger, QredoRendezvousAuthenticationType) {
     QredoRendezvousAuthenticationTypeAnonymous = 0,                         // The tag is just a string, has no cryptographic identity
